fix: APNS.CertConfig moved back to APNS.Config
Required for backwards compatability. Config parsing is now handled in
APNS.ConfigParser, which returns the appropriate struct for the given
opts.

`jwt_` prefix has been removed from the various attributes on JWTConfig.
Still unsure whether or not to remove it for static Mix.Config options.

TODO: Cleanup documentation and tests.
