+++ lib/pigeon/apns/config.ex
@@ -1,55 +1,57 @@
 defmodule Pigeon.APNS.Config do
-  @moduledoc """
+  @moduledoc ~S"""
   Configuration for APNS Workers
   """
 
-  alias Pigeon.APNS.{CertConfig, JWTConfig}
-
-  @apns_production_api_uri "api.push.apple.com"
-  @apns_development_api_uri "api.development.push.apple.com"
-
-  @type t :: CertConfig.t() | JWTConfig.t()
+  defstruct name: nil,
+            reconnect: true,
+            cert: nil,
+            certfile: nil,
+            key: nil,
+            keyfile: nil,
+            uri: nil,
+            port: 443,
+            ping_period: 600_000
 
   @typedoc ~S"""
-  Options for configuring APNS connections.
+  APNS configuration struct
 
-  - `:name` - Registered worker name.
-  - `:mode` - If set to `:dev` or `:prod`, will set the appropriate `:uri`
-  - `:cert` - Push certificate. Can be one of three options:
-    - Static file path
-    - Full-text string of the file contents (useful for environment variables)
-    - `{:my_app, "certs/cert.pem"}` (indicates path relative to the `priv`
-      folder of the given application)
-  - `:key` - Push private key. Same as `:cert`
-  - `:uri` - Push server uri. If set, overrides uri defined by `:mode`.
-    Useful for test environments.
-  - `:port` - Push server port. Can be any value, but APNS only accepts
-    `443` and `2197`
-  - `:ping_period` - Interval between server pings. Necessary to keep long
-    running APNS connections alive. Defaults to 10 minutes.
-  - `:jwt_key` - 
-  - `:jwt_key_identifier` - 
-  - `:jwt_team_id` - 
+  This struct should not be set directly. Instead use `new/1`
+  with `t:config_opts/0`.
+
+  ## Examples
+
+      %Pigeon.APNS.Config{
+        name: :apns_default,
+        reconnect: true,
+        cert: nil,
+        certfile: "cert.pem",
+        key: nil,
+        keyfile: "key.pem",
+        uri: "api.push.apple.com",
+        port: 443,
+        ping_period: 600_000
+      }
   """
-  @type config_opts :: [
+  @type t :: %__MODULE__{
           name: atom | nil,
-          mode: :dev | :prod | nil,
-          cert: binary | {atom, binary},
-          key: binary | {atom, binary},
           reconnect: boolean,
-          ping_period: pos_integer,
+          cert: binary | nil,
+          certfile: binary | nil,
+          key: binary | nil,
+          keyfile: binary | nil,
+          uri: binary | nil,
           port: pos_integer,
-          uri: binary,
-          jwt_key: binary | {atom, binary},
-          jwt_key_identifier: binary | nil,
-          jwt_team_id: binary | nil
-        ]
+          ping_period: pos_integer
+        }
+
+  alias Pigeon.APNS.ConfigParser
 
   @doc false
   def default_name, do: :apns_default
 
   @doc ~S"""
-  Returns a new `APNS.CertConfig` or `APNS.JWTConfig` with given `opts` or name.
+  Returns a new `APNS.Config` or `APNS.JWTConfig` with given `opts` or name.
 
   If given an atom, returns the config specified in your `mix.exs`.
 
@@ -63,108 +65,158 @@ defmodule Pigeon.APNS.Config do
       ...>   port: 2197,
       ...>   ping_period: 300_000
       ...> )
-      %Pigeon.APNS.CertConfig{uri: "api.push.apple.com", name: :test,
+      %Pigeon.APNS.Config{uri: "api.push.apple.com", name: :test,
       ping_period: 300000, port: 2197, reconnect: false}
 
       iex> config = Pigeon.APNS.Config.new(:apns_default)
-      iex> %{config | certfile: nil, keyfile: nil, jwt_key: nil, jwt_keyfile: nil, jwt_key_identifier: nil, jwt_team_id: nil} # Hide for testing
-      iex> match? %_{uri: "api.development.push.apple.com", name: :apns_default, ping_period: 600_000, port: 443, reconnect: false}, config
+      iex> %{config | certfile: nil, keyfile: nil} # Hide for testing
+      iex> match? %_{uri: "api.development.push.apple.com",
+      ...> name: :apns_default, ping_period: 600_000, port: 443}, config
       true
   """
-  @spec new(atom | config_opts) :: CertConfig.t() | JWTConfig.t()
   def new(opts) when is_list(opts) do
-    case check_config(Enum.into(opts, %{})) do
-      :cert -> new_cert_config(opts)
-      :jwt -> new_jwt_config(opts)
-    end
-  end
-
-  def new(name) when is_atom(name) do
-    Application.get_env(:pigeon, :apns)[name]
-    |> Enum.to_list()
-    |> Keyword.put(:name, name)
-    |> new()
-  end
-
-  defp new_cert_config(opts) do
-    %CertConfig{
+    %__MODULE__{
       name: opts[:name],
       reconnect: Keyword.get(opts, :reconnect, false),
-      cert: cert(opts[:cert]),
-      certfile: file_path(opts[:cert]),
-      key: key(opts[:key]),
-      keyfile: file_path(opts[:key]),
-      uri: Keyword.get(opts, :uri, uri_for_mode(opts[:mode])),
+      cert: ConfigParser.cert(opts[:cert]),
+      certfile: ConfigParser.file_path(opts[:cert]),
+      key: ConfigParser.key(opts[:key]),
+      keyfile: ConfigParser.file_path(opts[:key]),
+      uri: Keyword.get(opts, :uri, ConfigParser.uri_for_mode(opts[:mode])),
       port: Keyword.get(opts, :port, 443),
       ping_period: Keyword.get(opts, :ping_period, 600_000)
     }
   end
 
-  defp new_jwt_config(opts) do
-    %JWTConfig{
-      name: opts[:name],
-      reconnect: Keyword.get(opts, :reconnect, false),
-      uri: Keyword.get(opts, :uri, uri_for_mode(opts[:mode])),
-      port: Keyword.get(opts, :port, 443),
-      ping_period: Keyword.get(opts, :ping_period, 600_000),
-      jwt_key: opts[:jwt_key],
-      jwt_keyfile: file_path(opts[:jwt_key]),
-      jwt_key_identifier: Keyword.get(opts, :jwt_key_identifier),
-      jwt_team_id: Keyword.get(opts, :jwt_team_id)
-    }
-  end
+  def new(name) when is_atom(name), do: ConfigParser.parse(name)
+end
 
-  defp check_config(%{cert: cert, jwt_key: jwt_key})
-       when not is_nil(cert) and not is_nil(jwt_key) do
-    raise Pigeon.ConfigError,
-      message:
-        "Invalid Configuration. :cert and :jwt_key can't both be configured."
-  end
+defimpl Pigeon.Configurable, for: Pigeon.APNS.Config do
+  @moduledoc false
 
-  defp check_config(%{cert: _cert, jwt_key: nil}), do: :cert
-  defp check_config(%{cert: cert}) when not is_nil(cert), do: :cert
-  defp check_config(%{cert: nil, jwt_key: _jwt_key}), do: :jwt
-  defp check_config(%{jwt_key: jwt_key}) when not is_nil(jwt_key), do: :jwt
+  import Pigeon.Tasks, only: [process_on_response: 2]
 
-  defp check_config(_) do
-    raise Pigeon.ConfigError,
-      message:
-        "Invalid Configuration. Either :cert or :jwt_key must be configured."
-  end
+  alias Pigeon.APNS.{Config, Notification, Error}
 
-  defp uri_for_mode(:dev), do: @apns_development_api_uri
-  defp uri_for_mode(:prod), do: @apns_production_api_uri
-  defp uri_for_mode(_else), do: nil
+  @type headers :: [{binary(), any()}]
 
-  @doc false
-  def file_path(nil), do: nil
+  @type sock :: {:sslsocket, any, pid | {any, any}}
 
-  def file_path(path) when is_binary(path) do
-    if :filelib.is_file(path), do: Path.expand(path), else: nil
-  end
+  # Configurable Callbacks
 
-  def file_path({app_name, path}) when is_atom(app_name),
-    do: Path.expand(path, :code.priv_dir(app_name))
+  @spec worker_name(any) :: atom | nil
+  def worker_name(%{name: name}), do: name
 
-  @doc false
-  def cert({_app_name, _path}), do: nil
-  def cert(nil), do: nil
+  @spec max_demand(any) :: non_neg_integer
+  def max_demand(_config), do: 1_000
 
-  def cert(bin) do
-    case :public_key.pem_decode(bin) do
-      [{:Certificate, cert, _}] -> cert
-      _ -> nil
+  @spec connect(any) :: {:ok, sock} | {:error, String.t()}
+  def connect(%{uri: uri} = config) do
+    uri = to_charlist(uri)
+
+    case connect_socket_options(config) do
+      {:ok, options} ->
+        Pigeon.Http2.Client.default().connect(uri, :https, options)
+
+      error ->
+        error
     end
   end
 
-  @doc false
-  def key({_app_name, _path}), do: nil
-  def key(nil), do: nil
+  @spec push_headers(Config.t(), Notification.t(), Keyword.t()) :: headers()
+  def push_headers(_config, notification, _opts) do
+    json = Poison.encode!(notification.payload)
 
-  def key(bin) do
-    case :public_key.pem_decode(bin) do
-      [{:RSAPrivateKey, key, _}] -> {:RSAPrivateKey, key}
-      _ -> nil
+    [
+      {":method", "POST"},
+      {":path", "/3/device/#{notification.device_token}"},
+      {"content-length", "#{byte_size(json)}"}
+    ]
+    |> put_apns_id(notification)
+    |> put_apns_topic(notification)
+  end
+
+  @spec push_payload(Config.t(), Notification.t(), Keyword.t()) ::
+          iodata | no_return
+  def push_payload(_config, notification, _opts) do
+    Poison.encode!(notification.payload)
+  end
+
+  def handle_end_stream(_config, stream, notification, on_response) do
+    %{headers: headers, body: body, status: status} = stream
+
+    case status do
+      200 ->
+        n = %{notification | id: get_apns_id(headers), response: :success}
+        process_on_response(on_response, n)
+
+      _error ->
+        reason = Error.parse(body)
+        Error.log(reason, notification)
+        notification = %{notification | response: reason}
+        process_on_response(on_response, notification)
     end
   end
+
+  @spec schedule_ping(any) :: no_return
+  def schedule_ping(%{ping_period: ping}) do
+    Process.send_after(self(), :ping, ping)
+  end
+
+  def close(_config) do
+  end
+
+  def put_apns_id(headers, notification) do
+    case notification.id do
+      nil -> headers
+      id -> headers ++ [{"apns-id", id}]
+    end
+  end
+
+  def put_apns_topic(headers, notification) do
+    case notification.topic do
+      nil -> headers
+      topic -> headers ++ [{"apns-topic", topic}]
+    end
+  end
+
+  def get_apns_id(headers) do
+    case Enum.find(headers, fn {key, _val} -> key == "apns-id" end) do
+      {"apns-id", id} -> id
+      nil -> nil
+    end
+  end
+
+  def connect_socket_options(%{cert: nil, certfile: nil}) do
+    {:error, :invalid_config}
+  end
+
+  def connect_socket_options(%{key: nil, keyfile: nil}) do
+    {:error, :invalid_config}
+  end
+
+  def connect_socket_options(config) do
+    options =
+      [
+        cert_option(config),
+        key_option(config),
+        {:password, ''},
+        {:packet, 0},
+        {:reuseaddr, true},
+        {:active, true},
+        :binary
+      ]
+      |> add_port(config)
+
+    {:ok, options}
+  end
+
+  def cert_option(%{cert: cert, certfile: nil}), do: {:cert, cert}
+  def cert_option(%{cert: nil, certfile: file}), do: {:certfile, file}
+
+  def key_option(%{key: key, keyfile: nil}), do: {:key, key}
+  def key_option(%{key: nil, keyfile: file}), do: {:keyfile, file}
+
+  def add_port(opts, %{port: 443}), do: opts
+  def add_port(opts, %{port: port}), do: [{:port, port} | opts]
 end
